requirements.txt
aerich==0.9.2
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgi-lifespan==2.1.0
asyncclick==8.3.0.7
asyncpg==0.30.0
attrs==25.4.0
bcrypt==5.0.0
black==25.11.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
coverage==7.11.3
cryptography==46.0.3
dictdiffer==0.9.0
ecdsa==0.19.1
fastapi==0.121.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
iso8601==2.1.0
jose==1.0.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pypika-tortoise==0.6.2
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
tomli_w==1.2.0
tortoise-orm==0.25.1
trio==0.32.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
